requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-jose[cryptography]
passlib[bcrypt]
pymongo
motor
python-dotenv
aiofiles
email-validator
cloudinary